--- amazon/tests/test_amzn_views.py
@@ -1,6 +1,0 @@
-from amazon.tests.test_amzn_models import *
-
-
-class Test_SpapiBase():
-    pass
-
