+++ amazon/tests/test_views.py
@@ -1,0 +1,6 @@
+from amazon.tests.test_models import *
+
+
+class Test_SpapiBase():
+    pass
+
