--- dashboard/tests/test_base_credentials.py
@@ -1,4 +1,0 @@
-from django.test import TestCase
-from dashboard.d_models import BaseCredential
-from dashboard.tests.test_store import TestStoreProfile
-
