+++ dashboard/tests/test_views.py
@@ -1,0 +1,8 @@
+from django.test import TestCase
+from dashboard.d_views import Dashboard
+
+
+class TestDashboard(TestCase):
+    def setUp(self):
+        self.dash_client = Dashboard()
+
