+++ static/css/common.css
@@ -28,15 +28,15 @@ a {
 
 
 /* Store detail */
+.store-detail{
+    display: flex;
+    gap: 10px;
+}
+
 .order,.report {
     border: 1px solid black;
     border-radius: 3px;
     padding: 10px;
-}
-
-.store-detail{
-    display: flex;
-    gap: 5px;
 }
 
 .date-range{
